Fix third-party license file generation.
Replace the license:aggregate-add-third-party goal in the parent pom
by a license:add-third-party goal in the distribution pom. The 
license:aggregate-add-third-party goal fails if the local maven 
repository is empty because it is executed before the install phase,
so before org.elasticlib dependencies are available. Generating the 
third-party license file in the distribution module fixes that.

Besides that, only a single THIRD-PARTY.properties file in the 
distribution module is needed in this configuration, instead of having
to maintain one per module.

Warns are no longer triggered during build because of forked 
license:add-third-party goals skipped on a pom packaged module.

Also exclude org.elasticlib dependencies and test-scoped dependencies
from the generated license file.